+++ platformio/managers/package.py
@@ -27,6 +27,7 @@ from platformio.downloader import FileDownloader
 from platformio.unpacker import FileUnpacker
 from platformio.vcsclient import VCSClientFactory
 
+
 # pylint: disable=too-many-arguments